User manual: Small text updates and new screenshots
This adds Gnome-based screenshots to the manual for most of the images not
involving the profile. More of the images remain to be done.

I rewrote small sections of the text to reflect some of the recent
changes. Several images still need to be replaced. The profile images will
be done once the profile panel has been finalised. Some text also still
needs working.

Signed-off-by: Willem Ferguson <willem@willem-Precision-M4700.(none)>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>