User manual: replace some images, minor updates
Except for the images involving the dive profile, all the images now are
from Gnome 3. Images for dive profile awaits finalisation of the new dive
profile.

Small changes to the text of the manual have been done, the most important
of which is that dealing with using the unified import dialogue of
Subsurface.

Signed-off-by: Willem Ferguson <willem@willem-Precision-M4700.(none)>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>