Make sure that there are plot_info entries for events
We use the plot_info as basis for our dataModel in the new profile. In
order to be able to place events at the right spot we need to have a
plot_info entry when the event happens. So let's add interpolated entries
not only based on time but also whenever there's an event between them.

This should address Robert's comment in commit 0474fe70fc47 ("New profile:
add image pixmaps for image events").

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>