Fix event saving in git format due to wrong argument order
Oops.  Really stupid bug in event saving, resulting in bad event lines,
that I didn't notice until I started trying to parse them.

The argument order is a bit mixed up, which is partly why this happened.
But considering that this is the worst bug I've hit so far in the saving
code, I guess I shouldn't complain too much.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>